pyproject.toml
[tool.poetry]
name = "stagvault"
version = "0.1.0"
description = "Accessible media database for vector files, images, audio, and textures"
authors = ["Michael Ikemann"]
license = "MIT"
readme = "README.md"
packages = [{include = "stagvault"}]

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.10"
pyyaml = "^6.0"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
httpx = "^0.28"
click = "^8.1"
rich = "^13.9"
aiofiles = "^24.1"
pillow = "^11.0"
resvg-py = "^0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^6.0"
pytest-asyncio = "^0.25"
pytest-playwright = "^0.6"
ruff = "^0.9"
mypy = "^1.14"
types-pyyaml = "^6.0"
types-aiofiles = "^24.1"

[tool.poetry.scripts]
stagvault = "stagvault.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
